Update release notes with SAND number
The release notes have been R&Aed.

We have not been good about sticking the SAND number into this
document, so I made sure to do so this time.  It's just plain text at
the very end.

I also updated the comments about R&A with the sand number in them.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/89