Fix CMake so "Xyce_VERSION_EXTRA" is added to install name
Our Xyce install directories are supposed to be different for Full,
NoRad, and OpenSource builds (which these days should properly be
called XyceRad, XyceNF, and Xyce).

Our old CMake used to stick "NORAD" or "OPENSOURCE" into the directory
name as appropriate, but this got overlooked.  I have put the
necessary stuff back in.  Xyce_VERSION_EXTRA should be appended to
Xyce_VERSION_STRING_LONG if it's defined and not null.