Add support for HERTZ in data statements
Issue #93

Also, SON bugilla issue 50

I previously added support for the Hspice special variables TEMPER and
HERTZ, which mean the same things as TEMP and FREQ in Xyce.

Most of the rest of the Xyce source code doesn't need to know about
this - the expression library treats them as synonyms and then requests
the state variable for frequency either way.

But, there was one exception that I found, and that was DATA statements.
Part of the point of DATA statements was to give another way of sweeping
(or modifying) the frequency.  This commit adds HERTZ as an allowed
synonym for FREQ in the AC and NOISE classes in data tables.