Adds helper functions to create a N_PDS_ParMaps or N_PDS_Comm
Most constructors that generate N_PDS_ParMaps are agnostic to the
linear algebra implementation, so they can be placed in helper functions
to further separate the object interface from the implementation.
Another method was added to the help determine which N_PDS_Comm needs
to be used in the multivector.  That logic was duplicated in several
places, so it's best to simplify it to one helper function.

The Linear::Builder objects were also creating these ParMaps using
a method in the Parallel::Manager that was just calling the ParMaps
constructor.  Now the Linear::Builder calls the helper function
directly and the method can be removed from the Parallel::Manager.

Gitlab issue #36