Implement "option scale" (bug 1041 SON)
This commit implements the "options scale" feature of Hspice/ngspice.

In Xyce, the netlist command is:  .options parser scale=1.0e-6

The only other parameter that can be specified in ".options parser" is
the model_binning=true/false flag.

For cases when model binning is used, the value for "scale" needs to be
know in the model binning code, which is why this is a parser parameter
and not a "options device" parameter.

I've modified the model binning code and I have also modified all the
hand-written MOSFET models.   I have not touched any rad models and I
have not touched any ADMS models.  MOSFET models in those directories
should also get modified to fully close this bug.

Currently, this code is a little fragile, but it doesn't break any
tests.  Mainly, I haven't decided what to do about "option scale" when a
device is relying on length/area parameters that come from the "options
device" specification, or from the .MODEL, rather than from the
device instance line.

Currently, the scale parameter is ONLY applied to instance-level sizing
parameters L,W,AS,AD,PS,PD.  It is only applied to parameters that were
"given" by the user.

It is not applied to their model-level equivalents, or to their "options
device" equivalents.

The vast majority of the time, MOSFETS are specified with sizing
information on the instance line, so this issue is definitely a corner
case.   But we need to decide what to do there.