Added missing variable to the second constructor
In my work on bug 1041 SON, to add .option parser scale=#, I added a new
variable "lengthScale_" to this class.

I initialized it in the first of 2 constructors.  I neglected to
initialize it in the second one.  This commit fixes that mistake.

I do not know if this is the reason for test failures on some of the
continuous integration platforms, but fixing it is the right thing to
do.