Add SAND #'s to 7.2 Ref/User Guide Entries in circuit.bib
This commit adds the SAND Report numbers for the Xyce 7.2 Users'
Guide and Reference Guide to the circuit.bib file on the develop
branch.  It does not appear that was done on the release branch.
So, I'm making this change into a standalone commit, as I update
those two guides for the start of the Xyce 7.3 Release cycle.