Remove pointless "hello world" from unit test dir
Back during the spring I decided to start applying coverage testing via
gcov to my expression library unit tests.  Early in the process, to
teach and/or remind myself about the details of using gcov, I first
applied it to the simplest program I could think of, "hello world".
I didn't need to do that for very long and moved on quickly to the
actual programs for which I wanted to perform coverage tests.

But, I forgot that the "hello world" program (hello.C) was present in
this directory. It was also still in the gcov Makefiles for gcc8.
It should not have been commited, but I inadvertantly did so.

With this commit hello.C is removed, and it is also removed from those
two makefiles.