Separates parallel map object into abstract interface and Epetra implementation
The is the next large step in separating the fundamental parallel distribution
and linear algebra classes into abstract interfaces and implementations.  The
N_PDS_ParMap is an object that is mostly used in the linear algebra services,
but has been a thin wrapper around an Epetra_Map object.

The N_PDS_ParMap class is now an abstract class, with no mention of linear
algebra implementation.  The implementation is in the inherited class,
N_PDS_EpetraParMap.  So, the base class is used in the linear algebra
interfaces and cast back to an Epetra-based object whenever necessary in
the linear algebra implementations.  The N_PDS_ParMap interface has been
extended to reduce the need to make that cast until actual computations
with Epetra objects are necessary.

Gitlab issue #36