Begin adding "tablefile" capability to tableOp
Issue #101.

There has been a long-running issue to change how filenames for
time-value tables are specified.  In the old way, they would be
specified as param="filename".  This was problematic because in many
examples, people expect something with quotes around it to be treated as
a string and not automatically be assumed to be a filename.

This concern was reported a long time ago in the SON bugzilla issue 274,
and was partially fixed a while back with the keyword "tablefile".

That fix was only partial, however, as the expression library (both old
and new) was never adapted to understand the keyword "tablefile", so
using this capability with expressions often breaks.  Part of the
problem was that the old expression library was klunky enough that
nobody wanted to implement this capability directly inside of it, so an
awkward workaround was set up in the IO parser files.

We decided that the best way to fix this was to simply make it so that
all the "tablefile" processing was simply done by the expression
library.   That will make the code for this a lot cleaner and will cover
all the expression-basd use cases.

This commit is a preliminary commit, that adds tokens and parser rules
for "tablefile".  They now call a new constructor in the tableOp class,
that has as its only argument the std::string of the filename containing
the table.

To do: I have not yet fleshed out this constructor to do anything with
this filename.  That will come in future commits, along with unit tests.