Support for "tablefile" in expr library
Issue #101.

This commit completes preliminary implementation of the tablefile
capability in the expression library, complete with two very small unit
tests.  More of these will come later.

I had a previous commit on this topic, in which the lexing and parsing
mechanics for tablefile were added.  This commit fleshes out the new
tableOp constructor, which takes as its only argument the string
name of the file.

At the moment, this is hardwired to assume that the input varible
(x-axis) is time.  The more general, non-file "table" capability
allows you to set this to anything you like.  It doesn't have to be
time.

So, that is probably an easy enhancement that can be made.

Also, Rich pointed out that a reasonable syntax for this could also just
be table("filename") instead of tablefile("filename").  I may add this
as an alternative syntax later.

But for now, this is in the expression code and seems to work.

Next up: refactoring the relevant IO package function(s) so that they do
NOT attempt to parser "tablefile" themselves.