Finish tablefile capability, incl case-sensitivity
Issue # 101.

In my previous commit, I managed to implement a preliminary version of
the tablefile capability, which allowed the tableOp class read data
directly from a file, rather than have the parser do that work, and then
convert the data table to an expression string, to be re-parsed again by
the expression library.  That approach was hackish and included
unecessary work.

In this commit, I am refining things a bit further, and I think I mostly
have it done.  This started with me commenting out most of the guts
of two functions in the circuit context class, and replacing those guts
with a much simpler string that used "tablefile" and the filename.

I discovered when I did this that most of the test suite passed but at
least one thing failed because of case-sensitivity.  So this required
that I got back and rewrite a little bit of the expression library.

The issue was this.  The usage of Flex/Bison for the new expression
library is configured to be case-insensitive.  However, it doesn't
actually change the case of anything to accomplish this.   This became a
small problem b/c in the Xyce IO package, things like parameter names,
function names, etc are all up-cased to render them case-insensitive.
So, for the purposes of interacting with the rest of Xyce, I set up the
expression library to upcase the entire expression string prior to
parsing it.  That way, param names, etc would always match the rest of
Xyce for searches, etc.

However, if the expression string now contains potentially strings
corresponding to filenames, I can't up-case the whole thing.  Filenames
have to maintain the correct case to work on Linux and other
case-sensitive file systems.

So, in order to fix this, I no longer up-case the entire strings.
Instead, I upcase some tokens, such as parameter names, during the
lexing phase.  That way, just the things that Xyce expects to be upper
case are modified, and everything else is left alone.

The remaining thing here is that I need to make extra sure that the
tokenizing of the filename in the lexer covers all the file name
characters that are allowed on a typical file system.  I think I
don't quite have this right at the moment.  Close, but not quite.  So
that will take a little more refinement.