Removes N_LAS_LAFactory and initial value from vector/matrix constructors
The N_LAS_LAFactory, which has been around for some time now, is just
a thin layer over the linear algebra classes that create matrices and
vectors.  This is, likewise, the job of the builder classes for each of
the analysis types.  The builders mostly call the linear algebra
constructors directly, except for a few cases.  To avoid confusion
during this refactor, the N_LAS_LAFactory is being removed.

The builder classes create matrices and vectors, where the methods are
passed an initial value.  This initial value is ignored, and has been
for some time.  So, the creation methods have been cleaned up to remove
methods that aren't used and initial value arguments for ones that are
being used.

Lastly, the package tests in the <package_name>/test directories are
being removed.  They haven't been used for a long time and only cause
confusion during code refactors.

Gitlab issue #36