Remove deprecated string(" ") syntax
When double quotes were used to signify a filename
from which data should be read and converted to TABLE, there
wasn't a way for a parameter to be just a string.  While
the tablefile("") syntax was introduced as a way to import
a file into a TABLE, string("") was used to specify a string
valued parameter.  With the removal of the deprecated double
quote to TABLE support, the string("") syntax can be removed.
Now double quoted strings are handled as just that.  A
double quoted string

Issue#101