Make a small optimization to transmission line interpolation
The transmission line does extrapolation if the history is too old to
do interpolation (which only happens when the last time step is larger
than the transmission line delay).

One can detect this condition trivially by looking at the last element
of the history, and if it is true, there is no point in searching the
list with "lower_bound" to find out.  That is just wasted search.

This commit skips the search when it is pointless.