Handle Rise/Fall/Cross Vals < -1 for Extrema Measures
The RISE, FALL and CROSS qualifiers were implemented, at the
request of an internal customer, for the Extrema measures (MAX, MIN
and PP).  Those three qualifiers are not supported for MAX, MIN and
PP measures in HSPICE.  So, I decided to have RISE/FALL/CROSS values
less than -1 (or LAST) default to -1, with a warning message.

SON Bug 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313