Improve transmission line interpolation
The lossless transmission line, which has been in Xyce almost
unmodified since the very early days, works by keeping a history of
values as it accepts steps.  At each step, we look back in the history
to find what the other side of the line was doing one line delay in
the past.  It is almost never the case that we find exactly that time
in the history, and so must interpolate that history to get an
estimate of what the other side of the line would have been doing at
that time.

The Xyce lossless transmission line was based on the SPICE3F5
implementation as closely as I could back when I wrote it, in that it
always uses a three-point (quadratic) Lagrange interpolation to get at
the delayed values from history.

Unfortunately, trying to use this three-point formula when there is a
discontinuity in the history (as there will be when pulsed sources
are present) is wrong and produces some odd artifacts around the
delayed discontinuities (and by discontinuities I mean
discontinuities in derivative).  Instead of geting sharp corners that
match the behavior of the other side of the line, we get either odd
rounded corners or spiky overshoots.

This commit fixes several problems that have recently been observed
when the detailed behavior of the transmission line has been
inspected.

The transmission line is already scanning for discontinuities in
history so it can set breakpoints --- I now use the same test to
determine that we should not do quadratic interpolation, and do linear
interpolation using only two points of history (either straddling the
desired point or immediately before the desired point when we're
extrapolating).

With this fix, two test cases started failing --- precisely because
the bizarre interpolation artifacts present in their gold standards no
longer exist.  I will be committing changed gold standards for those
two tests shortly (and have presented an argument for why this is a
reasonable change to the test in the associated issue report).

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/103