Quick hack to try w/no breakpoints in expr table
Issue #107.

This commit implements a quick hack to try out excluding breakpoints
from an expression-based table.

This quick hack basically pre-pends the word "fast" to "table" and
"tablefile" and this results in a flag being set in the tableOp class to
disable breakpoints.

This makes big tables run much faster, unsurprisingly.  But of course,
doing it can be risky, if the table represents discontinuous data.