Interim commit for Issue #107 and Issue #80
I have been working on refactory how tables are handled.  Part of that
refactoring is to add more capabilities and part of it has been to make
the parsing work better with fewer lexing and parsing rules.

This commit is one step in the direction of better parsing rules.  it
doesn't add any new capabilties, but refactors some of the table-based
lexing and parsing so that now "fasttable" and "table" can be handled by
the same rules.  This type of approach makes the grammar much more
manageable.