More table/spline refactors
Issue #107 and Issue #80.

I have been working towards several synergistic goals.

(1) combining as many parser rules as possible.  This avoids a "rule
explosion" when lots of new table options are added, as I am in the
process of doing.

(2) handling ALL types of interpolation in the N_UTL_Interpolators
class, including PWL and splines in a single tableOp class.

(3) making all of these interpolators available to the user as options
via the expression parser.

(4) making breakpoints in the table optional.  Currently, my thinking
is that they are only an option for a PWL table, and are hardwired to be
OFF for any higher-order interpolation, as that should be smooth anyway.
Also, this must be a user-option, at least for PWL tables.

(5) making file-based tables work as cleanly as possible.  Currently,
the expectation in Xyce is for tables from files to be specified as
"tablefile", but my goal is to make table("filename") allowed as well.
If this is done, then this should also be easy: spline("filename").

Several of my recent commits have been in this direction.  This commit
is a huge step towards accomplishing this goal.  The tableOp class is
now nearly done, with respect to the above issues.    It is now set up
to use ALL of the intepolators in the N_UTL_Interpolators class, which I
wrote a number of years ago and have periodically revised.

The expression parser still needs some work, as it only supports "spline".
I also want it to support "akima", "cubic", "BLI" and "wodicka".

BLI = Barycentric Lagrange Interpolation.

Anway, this doing pretty well.