Parser support for all interpolators
Issue #107 and Issue #80.

In my previous commits, I modified the tableOp class to support using
all the different interpolators available in N_UTL_Interpolators.h.
However, I had only partially supported them in the parser.

With this commit, they are all supported in the parser as well.  I have
also managed to combined a lot of the "table" and "spline" rules, which
makes the grammar a little less complex.