Combining more rules.
Issue #107, Issue #101 and Issue #80.

This combines more table-related rules, to make the grammar a little
smaller.

This is the final comit (for now) on this topic.  To summarize, the
work accomplished over all the commits includes the following:

(1) I have made it possible for the TABLE operator to run without
breakpoints.  This is enabled by prefixing the string "FAST" to either
"TABLE" or "TABLEFILE".  ie, "FASTTABLE" and "FASTTABLEFILE".

(2) I have made it possible for users to use various interpolation
functions as an alternative to the linear PWL interpolation performed by
TABLE.  It is now possible to use all of the interpolators in the
N_UTL_Interpolators.h file.  So, the following keywords work in place of
TABLE, and use all the same syntax as TABLE:  SPLINE, AKIMA, CUBIC,
WODICKA and BLI.   SPLINE and AKIMA are synonym and will apply Akima
splines.  CUBIC is for cubic splines, WODICKA is for wodicka splines and
BLI is for Barycentric Lagrange Interpolation.

I chose to have SPLINE be a synonym with Akima, rather than Cubic,
because it tends to be better behaved numerically.

Note, BLI does not scale well (in the current implementation at least) for
really large data sets.  But all the splines do scale pretty well with
large data sets.

(3) The various spline functions do NOT produce breakpoints as it is
assumed they will produces smooth waveforms.

(4) the various spline functions can all have the word "FILE" added as
a suffix to their keywords.  In other words:  SPLINE -> SPLINEFILE,
WODICKA -> WODICKAFILE, etc will work.

(5) Finally, to read tables from a file, it is not necessary to to
append the "FILE" suffix to the keywrod.  The following all work as
well:  TABLE("filename"), SPLINE("filename"), AKIMA("filename"),
CUBIC("filename"), WODICKA("filename") and BLI("filename).

(6) All of this is handled in a single ast.h class: tableOp.  I had
previously considered separate operator  classes for all the options,
above, but decided this was simpler and easier to maintain.

(7) similarly, I was able to simplify the TABLE grammar a great deal, so
now the same lexer/parser rules are used for all of the above.  I had
prevously created separate rules for different variations, and that
wasn't going to scale very well.  Now, the rules are much more
compressed.  Part of this was made possible by having all of the above
capability handled by the same tableOp class.