Removes unnecessary use of "petra" naming and more Epetra headers
The name "petra" was added to functionality in linear algebra objects
where it was unnecessary.  For instance, "printPetraObject" is a
popular method in several linear algebra classes.  Really the only
task the method is being asked to perform is to "print" the object,
whether it is a "petra" based object or not.  Anyways, the naming
is unnecessary to convey the intent, so these names are being
removed for simpler names to enable interface abstraction.  Replacing
these methods with the new name touched a lot of files.

While searching for "petra", more Epetra headers were found that
are no longer necessary.

Gitlab issue #36