Fix valgrind errors introduced by table refactor
With the new handling of TABLE and other related features such as SPLINE, I had
to rearrange how various arguments were handled in the lexer and parser.  In
doing so I neglected to include a few delete commands.  That caused
several valgrind errors in nightly testing.

This commit should fix it.