Add an ideal delay device
Issue #102 on gitlab-ex and the long-closed-as-"wontfix" issue 165 on
joseki-vm called for implementation of an ideal delay device without
the impedance matching issues of the lossless transmission line.

This commit provides such a device.

The device is used just like a unity gain voltage-controlled voltage
source, only the output of the source is the time-delayed value of the
controlling voltage (V(pos,neg) at time t is V(control+,control-) at
time t-TD).

A required instance parameter TD defines the delay.

An optional BPENABLED boolean instance parameter (default true) allows
enabling or disabling the device's ability to set time integrator
breakpoints when it detects a discontinuity on the input.  I have
found that disabling it makes the time integrator take many more steps
when the device has pulsed voltage source as input, but your mileage
may vary.

An optional EXTRAPOLATION boolean instance parameter (default true)
allows control of how the history lookup is done.  If true, and the
time integrator has take a time step longer than the device delay,
computation of the time-delayed value of the control voltage is
extrapolated from history.  If false, we use the current (unconverged)
value of the control voltage as an interpolation point and always
interpolate.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/102
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=165