Fix isnan/isinf use
A recent commit to N_IO_MeasureErrorFunctions.C added calls to
isnan/isinf.

All other files in Xyce that use this have a block of conditional code
meant to catch special portability issues with use of these functions.
This file lacked that.

It failed to build on my Ubuntu 16.04 LTS system with gcc 5.4 (the
default compiler on this COE system).

I have added the same portability code that exists in most other files
that use isnan/isinf, and the compilation problem went away.

Note that where isnan and isinf are bging used without this code,
they're being unconditionally called as "std::isnan" or "std::isinf",
and that would break on some systems, too (looking at Windows, where
this is often a problem).  But that very code is not even being built on
Windows, so I'm not worried yet.  We'll burn that bridge when we come to
it.

This sort of portability cruft probably ought to be centralized
somewhere, but I'm not going looking for a place to put it.