Add option to force linear interpolation in delay device
Ting would like to experiment with forcing the delay device to do only
linear interpolation.

Since the device is already written to do linear interpolation in some
cases, this was just a matter of adding a boolean instance parameter
to make it do so in all cases.  This parameter is "LINEARINTERP",
which defaults to false (so we do quadratic Lagrange interpolation if
the signal is not discontinuous, and linear only if it is
discontinuous), but if set to true will always do linear
interpolation.

Use of "EXTRAPOLATION" (default true) can change whether this linear
interpolation is allowed to be extrapolation or not.  If EXTRAPOLATION
is false, we use the current unconverged solution as one of the
interpolation points instead of extrapolating (this only happens if
the time step is larger than the delay time).

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/102