Remove unnecessary isnan/isinf cruft
Years and years ago, isinf and isnan were not part of the standard.

In those cases, we could not always call "std::isnan" or "std::isinf"
and had to find them in float.h as something weird.  On Windows it was
worse.

This is all now unnecessary, and all the cruft we had in configure to
detect isnan and isinf was just creating a HAVE_NAN_INF_SUPPORT that
was never used anywhere, and everywhere we *WERE* doing something
special, we were looking for HAVE__ISNAN_AND__FINITE_SUPPORT --- a
macro we apparently stopped defining ages ago (I think that went away
in the CMake refactor, which was unconditionally setting HAVE_ISNAN and
HAVE_ISINF to true --- neither of which are used anywhere, so I have
removed it).

isnan and isinf are part of the C++11 standard and are always
available as std::isnan and std::isinf.  We were already using them
this way unconditionally in a few places.

This commit removes the now-useless cruft in configure.ac and config
that test for isnan/isinf, and replaces all use of isnan() and isinf()
macros (that in reality always turned out to be defined as
std::isnan() and std::isinf()) with std::isnan() and std::isinf()
everywhere.

This not only fixes the issue of my last commit, but gets rid of
useless portability stuff that stopped being a problem long ago, and
isn't a problem now that we require C++11.