Fix for parallel hangs, Issue #109
Issue # 109

This was observed by Pete today and there was much subsequent
discussion.

To recap:

On some platforms, a regression test would hang in parallel.  This is a
test that
(1) specified a table via the new "tablefile" keyword
(2) the file specified did not exist (was a directory)
(3) was supposed to cleanly exit with an error message.

The error message was emitted, but then followed by the code hanging.
This was because even though the file was not a valid file, the code
attempted to open it anyway.

All of this was happening because:
(1) ALL tablefile processing is now handled in the expression library.
The parser is no longer directly involved.
(2) in the expression library, for most errors (such as a "file not
valid" error) I have been using the yyerror function.
(3) the yyerror function simply calls the Xyce error handling functions,
but uses Xyce::Report::UserError().
(4) The "UserError" capability does not exit immediately.  It buffers
errors and only exits once it hits a barrier.
(5) it doesn't hit that barrier right away, and in the meantime attempts
to open the invalid file.

Fixing this was trivial.  It just required that I add and "else" to the
original if-statement in one of the tableOp class constructors.

Also, to be extra safe, I also added several if-statement tests to the
val(), dx() and other functions to avoid calling the interpolators in
the event that they don't contain data.  I don't think the lack of these
if-statements would have caused  problems before, but I did this to be
extra safe with it.