Implement part of spiceSinOp::dx() function
Issue #110.

This commit implements the dx() function in the spiceSinOp class (for
the SPICE_SIN capability).  The dx functions in the various spice
sources in the new expression library uncondintionally return 0.0, prior
to this commit.  Most uses of the spice sources don't require
derivatives, but fixing this issue is the right thing to do.

This commit only adds support for derivatives w.r.t. time.  Ultimately,
it should also handle derivatives w.r.t. other parameters, such as VA.

I added the time derivative so that this function would work correctly
with the computeBreakPoints function, which depends on obtaining
analytical time derivatives via dx.