Adds clone methods to vector and multivector interfaces
To eliminate the copy constructor from the N_LAS_Vector and
N_LAS_MultiVector class, clone methods have been created for
both classes.  It will also enable the code to distinguish
between a clone, which just creates a vector of the same shape,
and a cloneCopy, which also copies the values from the original
vector into the new vector.  There are many cases where a
copy of a vector/multivector is made and then zeroed out.
This is less than ideal because it's inefficient.

Classes that used the copy constructor have been updated to
use the clone methods instead.  Some logic in the analysis,
loaders, and nonlinear solver were changed to more efficiently
use clone and cloneCopy.

Gitlab issue #36