Modifications to speed up expression library
Issue #94
Issue #98

All of the changes in this commit are in the function
newExpression::getValuesFromGroup().

I have been running a customer circuit that Rich provided thru the
Instruments profiler.  This circuit has been running a lot slower in
Xyce 7.2 with the new expression library than it was in 7.1 with the old
library.

Part of the problem was that I had some mistakes in the
"LIMIT" operator.    Fixing those mistakes helped  (especially fixing
the limitOp::dx() function), but it didn't completely fix the
performance problem.

In the profiler, it looked as though too much time was being spent in
the getValuesFromGroup function, so I've modified it a bit to make it
faster.  It has helped but probably not enough to catch 7.1 yet.

The modifications are mostly about using the std::vector empty()
function to determine if certain function calls should be made, and also
getting rid of memory allocations that are not necessary.

I tagged Issue #94 and Issue #98 in this commit message.  Both of them
are about performance issues in the new expression library for modern
PDKs   (GF 45nm and GF 14nm).  I haven't actually proven that this
commit improves things for those PDKs yet, but it could.