Add breakpoints to the limitOp class
Issue #108

The limit operator is very similar to the "if" and "stp" operators, in
that it can produce a discontinuous signal depending on the input.  If
the input is dependent on time, it should produce breakpoints.

I had neglected to implement breakpoint related code for the limit
operator before.  This commit corrects that issue.

There is one small change to the computeBreakPoint function included in
this commit, which affects other classes that also call this funciton.

The computeBreakPoint function was doing a clear on the bpTimes_
vector argument, prior to performing a numerical calculation to
estimate a breakpoint.  Doing this didn't break anything for operators
that only call this function once, and only expect to find a single
breakpoint.

However, the limitOp class calls this function 2x in a row  (because
it has 2 comparison conditions rather than one), which means it was always
deleting the first breakpoint it found.  That was a mistake.  It is
fixed now.  The calling function is now responsible for clearing the
vector.  This meant I had to update the stpOp class and all the
comparison operator classes to accomodate this change.