Further fix for the limitOp
Issue #108

My earlier fix to the limitOp::dx() function was necessary but not quite
sufficient.  It broke 3 POWER_GRID tests.

The reason that it broke them is that the conditional that is performed
inside of the limitOp::val() and limitOp::dx() functions did not check
any of the arguments for Inf or Nan prior to performing the evaluation.

The Bsrc's in these circuits have arguments to LIMIT which can produce
divide-by-zero errors.  In most expressions this doesn't cause a
problem, because I trap and fix these issues.  But, until this commit, I
was only trapping for those issues in the "newExpression::evaluateFunction"
function.

But for operators such as limitOp, that fix comes too late.

Not doing the fixes inside of the operator results in the conditional making
the wrong choice.  And, for something like limitOp, a crazy Inf or Nan should
result in one of the limits being returned (and in the case of dx() that means
returning 0.0).

But if the conditional is making the wrong choice, it may return the
dx() of the subordinate operator instead of the dx() of a limit (which
is zero). The subordinate dx() might also be a crazy number under
these circumstances.

This commit moves the Xyce::Util::fixNan and Xyce::Util::fixInf
functions from the newExpression.h header into the ast.h header, and
then uses them in a few places.  I've applies them to the stpOp and the
limitOp, and part of the ifStatementOp.  The ifStatementOp fix is
incomplete,  however, as this fix really should be applied inside the
conditional ops.

However, I chose not to fix the conditional Ops in this commit.  The
conditional ops are based on a macro, and the macro isn't perfectly set
up to use the Xyce::Util::fixNan and Xyce::Util::fixInf functions.  So,
that will have to wait for a future commit.