Add error checking on a collapse problem
Xyce is somewhat particular about how we collapse nodes.  We can
collapse internal nodes onto each other by mapping node numbers.  We
can  collapse internal nodes onto external nodes easily, too.

Collapsing external nodes onto internal by mapping their node numbers
isn't OK.  It is also essential that the mapping be one-to-one, as
well.

Verilog-A allows authors to specify that nodes a and b should be collapsed by
saying:

  if (bias-independent-condition)
    I(a,b) <+ current;
  else
    V(a,b) <+ 0;

This is normally use to collapse a branch when only a resistor is
between them, when that resistance is zero.

It is very difficult to handle this correctly in Xyce/ADMS when
authors have too many nodes that collapse onto each other.  We have
some basic rules:

   if both nodes are internal:
     If the negative node is not already collapsing onto something else,
       collapse the negative onto the positive
     If the negative node already collapses onto something, but the
        positive node doesn't, then collapse positive onto negative

   if the positive node is external and negative internal
      collapse negative onto positive
   If negative node is external and positive internal,
      collapse positive onto negative

Some cases simply cannot be handled using so basic of a rule.
Verilog-A authors can write really nasty models with lots of
collapses, and are not particularly attentive to how they write them.
Apparently other compilers and codes aren't picky.  Sometimes it is
impossible to do what authors intend based on these rules and what
they actually wrote.

This is not foolproof, and some error checking was missing ---
"positive external negative internal" was unconditionally mapping the
internal node to the external without checking if there was some other
collapse of that internal node already.  This got caught in Jason's
attempt to implement the new MVSG_CMC model (issue #112).  It was
quietly letting us try to do a one-to-two mapping of a node.

This doesn't fix the MVSG_CMC model, but does cause it to emit a loud
fatal error when it tries to do a one-to-two mapping in one case.

MVSG_CMC can be fixed by a slight rewrite of three lines of collapse
code.  I've communicated those to Jason privately.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/112