Make flex/bison probes unconditional
Ever since the new expression library was added, flex and bison are no
longer optional.  The autotools build has always had the probe for
these tools inside a conditional that allowed them to be skipped if
the reaction parser were not requested.

The bad part of this is that any automake build that were done with
reaction parser off would have failed at compile time, because flex
and bison were not defined.

This commit removes that conditional and makes flex/bison required.
It leaves the reaction parser as optional, however, but changes its
conditional so that it is enabled if requested and disabled if it is
not requested, instead of depending on the existence of flex and
bison.

This completes the autotools part of issue #113.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/113