Apply fixNan and fixInf to comparison arguments
This is related to an issue I fixed on Friday in the limit operator and
the STP (step function) operator.

I recently fixed the limitOp::dx() function, but doing so caused a few
POWER_GRID regression tests to file, all of which used the limit
operator.   The reason that they broke was that the arguments to the
limit operator could sometimes evaluate to Nan and/or Inf.  The
expression library traps for Inf and Nan, and fixes them, but it was
only doing so right before returning from the
newExpression::evaluateFunction.    So, it wasn't bothering to fix these
things inside the expression tree.  It was only fixing them at the end.
At the time that I set this stuff up, it didn't occur to me that there
would be use cases where this would matter.

It turns out that for any operator performing a comparison, it matters.
I had created Xyce::Util::fixNan and Xyce::Util::fixInf functions
previously, so I applied them inside of the limitOp and the stpOp, as
well as parts of the ifStatmentOp.

But, I also needed to apply similar fixes to the comparison operators,
which are used by the ifStatementOp class.  Doing this was not possible
with the as-written Xyce::Util::fixNan and fixInf functions, because
they were too inlined.    So, today, I wrote new versions of those
functions, which made it easy to add this check to the comparison
operators.  They are now used by all the comparison ops, as of this
commit.