Make flex/bison required in CMake and some cleanup
With the addition of hte expression library, flex and bison are no
longer optional. This updates CMake to reflect that situation. It also
simplifies the use of the reaction parser flag. Before it was
enabled/disabled by the existance of flex/bison. Now it's just another
option in the "feature_modes.cmake" file.

I also cleaned up some of the other flex/bison setup in CMake (for
aesthetics, nothing functional should have changed).

Xyce/Xyce#113