Enable most Expression Library unit tests
This enables all possible Expression Library unit tests for a normal
build. The parserUnitTest tests require USE_TYPE_DOUBLE to be enabled.
However, that would break the rest of Xyce, so that is not included in
the build. Some comments were added to the CMakeLists.txt file to that
effect. Nevertheless, if "TEST_SUITE" is enabled in the build, all of
the other unit tests are built and made to be run with CTest.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/72